Move Matlab toolbox creation to separate script
